refactor(p2-shim): delegate browser filesystem descriptors
